mongo-spreadsheet-aat/requirements.txt
Brotli==1.0.9
click==8.0.3
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.1.0
Flask==2.0.2
Flask-Compress==1.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pandas==1.3.5
plotly==5.4.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
tenacity==8.0.1
Werkzeug==2.0.2
